requirements.txt
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
outcome==1.1.0
pip==21.1.2
pycparser==2.21
pyOpenSSL==21.0.0
python-dotenv==0.19.2
requests==2.26.0
selenium==4.1.0
setuptools==57.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.28
trio-websocket==0.9.2
trio==0.19.0
urllib3==1.26.7
wheel==0.36.2
wsproto==1.0.0
